### 2018-07-30 #862 @whedon
[REVIEW]: Pymer4: Connecting R and Python for Linear Mixed Modeling
**Submitting author:** @ejolly (<a href="http://orcid.org/0000-0002-6144-5789">Eshin Jolly</a>)
**Repository:** <a href="https://github.com/ejolly/pymer4" target ="_blank">https://github.com/ejolly/pymer4</a>
**Version:** 0.5.0
