### 2018-07-30 #862 @ejolly
Proof looks good to me. Thanks for agreeing to review all.
