### 2018-08-05 #862 @tomfaulkenberry
Hi @cMadan and @ejolly:

I am currently traveling with family for a little vacation before our Fall term starts.  When I get back to work later next week, I will immediately go through the paper and software and finish my review.  From what I've seen in my initial look, this shouldn't take long :)
