### 2018-08-05 ejolly/pymer4#28 @janfreyberg
[joss-review] Installation instructions
I just tried to install this and failed `brew install gcc` installed `gcc-8` rather than `gcc-7`.

Is it possible to fix the recommend bash command that sets the CC environment variable to use either?

Also, it may make sense to provide conda specific installation instructions: for users with no R installation, installing `rpy2` from conda-forge is likely the most frictionless way of installing R; by the looks of it, lmerTest and lsmeans are also on conda-forge.

If you're not so familiar with conda I can take a stab at that.
