### 2018-08-05 ejolly/pymer4#29 @janfreyberg
[joss-review] Paper issues
A few minor issues with the paper:

- [x] in the example, `get_resource_path` is never defined, I believe you're missing the import statement `from pymer4.utils import get_resource_path`
- [x] the line `model.plot('IV2')` from your paper throws a NotImplementedError
- [x] replace `langauge` with `language` in the first paragraph
- [x] references have no DOIs
