### 2018-08-05 ejolly/pymer4#30 @janfreyberg
[joss-review] Documentation issues
On the [Basic Usage](http://eshinjolly.com/pymer4/usage.html) page, `Lm` is not imported but later used; if someone simply steps through the code on that page, they receive errors.
