### 2018-08-05 #862 @janfreyberg
Hi @cMadan and @ejolly - just raised a few issues on the repo, but it's a well-documented, well-running package, so my issues are fairly minor:
https://github.com/ejolly/pymer4/issues/28
https://github.com/ejolly/pymer4/issues/29
https://github.com/ejolly/pymer4/issues/30

Nice work!
