### 2018-08-05 ejolly/pymer4#28 @ejolly
@janfreyberg Yup I can work on editing the environment variable command. I'd love some help on getting R and rpy2 as well as the r package dependencies from conda-forge if you already have experience setting that up during installation!
