### 2018-08-06 #862 @cMadan
@RMHogervorst, I am fine with you taking longer than two weeks if that would be more practical, but if you feel that reviewing this submission still is not feasible, then no worries about withdrawing as a reviewer. I do think, however, that if the package is too difficult for some to install, that is something we should try and improve in the review process, even if it means you focus on that one aspect rather than reviewing the submission as a whole (since we do have other reviews that can cover that side of things).

@tomfaulkenberry, not a problem, enjoy your vacation!

@janfreyberg, thanks for your work on the review so far!
