### 2018-08-06 ejolly/pymer4#28 @janfreyberg
Having just tested this, there seem to be dependency conflicts at the moment - rpy2 2.8.5 on conda forge depends on r=3.3.2, which at least on conda-forge is listed as incompatible with r-lmertest. Might depend on #25 being fixed or the r-lmertest conda-forge package being updated.
