### 2018-08-10 1b17c70
Fixed index naming loss when using rpy2==2.9.x
