### 2018-08-10 08e8929
Pymer4 should now work with rpy2>=2.9.x.
