### 2018-08-10 b7ae79a
Fixed documentation and paper issues from JOSS review.
