+++ paper.md
@@ -18,7 +18,7 @@ bibliography: paper.bib
 
 # Summary  
 
-In the social sciences the prevalence of clustered, correlated, and/or repeatedly measured data is very common. Modeling such data have classically involved the use of techniques such as the repeated measured analysis of variance (ANOVA) with correction factors applied to account for violations of model assumptions. More recently multi-level or hierarchical linear models have been brought to bear on modeling these kind of data given their flexibility. One of the most popular packages used for this purpose is the powerful `lme4` package in the R statistical programming langauge [@Bates:2014]. While this package is highly favored by individuals that utilize R, to date there exist few comparable packages in the scientific Python community. While some solutions do exist they are often more complicated to use and sometimes less flexible (e.g. `statsmodels` [@Seabold:2010] or require users to adopt a different statistical framework (e.g. Bayesian modeling, `bambi` [@Yarkoni:2016]). This leaves Python users in want for a tool that: a) is highly compatible with existing scientific python tools (e.g. pandas, `numpy`, `matplotlib`, `seaborn` [@McKinney:2012]), b) has an API that is easy to use but not unfamiliar to those who use to `lme4` , c) offers additional functionality that anticipates users' needs when analyzing real data (e.g. significance testing, simulating data, post-hoc analyses, etc).
+In the social sciences the prevalence of clustered, correlated, and/or repeatedly measured data is very common. Modeling such data have classically involved the use of techniques such as the repeated measured analysis of variance (ANOVA) with correction factors applied to account for violations of model assumptions. More recently multi-level or hierarchical linear models have been brought to bear on modeling these kind of data given their flexibility. One of the most popular packages used for this purpose is the powerful `lme4` package in the R statistical programming language [@Bates:2015]. While this package is highly favored by individuals that utilize R, to date there exist few comparable packages in the scientific Python community. While some solutions do exist they are often more complicated to use and sometimes less flexible (e.g. `statsmodels` [@Seabold:2010] or require users to adopt a different statistical framework (e.g. Bayesian modeling, `bambi` [@Yarkoni:2016]). This leaves Python users in want for a tool that: a) is highly compatible with existing scientific python tools (e.g. pandas, `numpy`, `matplotlib`, `seaborn` [@McKinney:2012]), b) has an API that is easy to use but not unfamiliar to those who use to `lme4` , c) offers additional functionality that anticipates users' needs when analyzing real data (e.g. significance testing, simulating data, post-hoc analyses, etc).
 
[…]
