+++ paper.bib
@@ -1,12 +1,14 @@
-@article{Bates:2014,
-  title={lme4: Linear mixed-effects models using Eigen and S4},
-  author={Bates, Douglas and Maechler, Martin and Bolker, Ben and Walker, Steven and others},
-  journal={R package version},
-  volume={1},
-  number={7},
-  pages={1--23},
-  year={2014}
-}
+@Article{Bates:2015
+   title = {Fitting Linear Mixed-Effects Models Using {lme4}},
+   author = {Douglas Bates and Martin M{\"a}chler and Ben Bolker and
+     Steve Walker},
+   journal = {Journal of Statistical Software},
+   year = {2015},
+   volume = {67},
+   number = {1},
+   pages = {1--48},
+   doi = {10.18637/jss.v067.i01}
+ }
 
 @inproceedings{Seabold:2010,
   title={Statsmodels: Econometric and statistical modeling with python},
@@ -15,14 +17,16 @@
   volume={57},
   pages={61},
   year={2010},
-  organization={SciPy society Austin}
+  organization={SciPy society Austin},
+  doi= {https://doi.org/10.5281/zenodo.275519}
 }
 
 @article{Yarkoni:2016,
   title={Bambi: A simple interface for fitting Bayesian mixed effects models},
   author={Yarkoni, Tal and Westfall, Jake},
   year={2016},
-  publisher={Open Science Framework}
+  publisher={Open Science Framework},
+  doi = {https://dx.doi.org/10.31219/osf.io/rv7sn}
 }
 
 @book{McKinney:2012,
@@ -47,6 +51,7 @@
   number={3},
   pages={90--95},
   year={2007},
+  doi = {10.1109/MCSE.2007.55},
   publisher={IEEE}
 }
 
@@ -54,7 +59,8 @@
   title={Seaborn: statistical data visualization},
   author={Waskom, Michael and Botvinnik, O and Hobson, P and Warmenhoven, J and Cole, JB and Halchenko, Y and Vanderplas, J and Hoyer, S and Villalba, S and Quintero, E and others},
   journal={URL: https://seaborn. pydata. org/(visited on 2017-05-15)},
-  year={2014}
+  year={2014},
+  doi = {https://doi.org/10.5281/zenodo.1313201}
 }
 
 @article{Kuznetsova:2017,
@@ -64,7 +70,8 @@
   volume={82},
   number={13},
   year={2017},
-  publisher={The Foundation for Open Access Statistics}
+  publisher={The Foundation for Open Access Statistics},
+  doi = {http://dx.doi.org/10.18637/jss.v082.i13}
 }
 
 @article{Lenth:2016,
@@ -74,7 +81,8 @@
   volume={69},
   number={1},
   pages={1--33},
-  year={2016}
+  year={2016},
+  doi={http://dx.doi.org/10.18637/jss.v069.i01}
 }
 
 @inproceedings{Huber:1967,
@@ -85,5 +93,6 @@
   number={1},
   pages={221--233},
   year={1967},
-  organization={University of California Press}
+  organization={University of California Press},
+  doi = {https://projecteuclid.org/euclid.bsmsp/1200512988}
 }
