### 2018-08-10 ejolly/pymer4#29 @ejolly
Ok these should all be fixed in the latest commit. I removed the `model.plot()` call since that specifically applies to continuous predictors only (for the time being). I've added DOIs for every source I could find but some don't appear to have any.
