### 2018-08-10 ejolly/pymer4#30 @ejolly
Fixed in latest commit on gh-pages branch.
