### 2018-08-10 ejolly/pymer4#28 @ejolly
So I made a few fixes that should make `pymer4` compatible with `rpy 2.9.x`. Any chance you can try again with the latest version from github? One installation issue I'm not certain about, and would be happy to receive any advice on, is the following:

The latest versions of `rpy2` (i.e. `>=2.9.x`) don't support Python 2. Therefore in order for `pymer4` to remain both Python 2 and 3 compatible, it seems like I have to freeze the requirements to an older version as I'm doing currently. However, if overall installation becomes easier by dropping Python 2 support and using `conda-forge` to install latest rpy2 and all R-package dependencies then maybe that's ultimately more convenient for users. Thoughts?
