### 2018-08-18 #862 @tomfaulkenberry
Overall, I found the paper to be concise, well-written, and perfectly describes to the reader why and how to use the software. Further, I found the repo documentation to be equally well-written.  I did not run into any issues with the software. There is ONE typo in the example code in the paper: on the line that begins

model = Lmer('DV'=

there is a missing parenthesis at the end of the line (it took me a while to figure out that this was the issue!).

Overall, well done on both the software and the paper!
