### 2018-08-20 553e20e
All code adheres to pep8 style guide now.
