+++ pymer4/models.py
@@ -11,7 +11,6 @@ import matplotlib.pyplot as plt
 from patsy import dmatrices
 import seaborn as sns
 import warnings
-from six import string_types
 from joblib import Parallel, delayed
 from pymer4.utils import (_sig_stars,
                           _chunk_boot_ols_coefs,
@@ -24,8 +23,9 @@ from pymer4.utils import (_sig_stars,
 __author__ = ['Eshin Jolly']
 __license__ = "MIT"
 
-warnings.simplefilter('always',UserWarning)
+warnings.simplefilter('always', UserWarning)
 pandas2ri.activate()
+
 
 class Lmer(object):
 
@@ -55,12 +55,14 @@ class Lmer(object):
 
     """
 
-    def __init__(self,formula,data,family='gaussian'):
+    def __init__(self, formula, data, family='gaussian'):
 
         self.family = family
-        implemented_fams = ['gaussian','binomial','gamma','inverse_gaussian','poisson']
+        implemented_fams = ['gaussian', 'binomial',
+                            'gamma', 'inverse_gaussian', 'poisson']
         if self.family not in implemented_fams:
-            raise ValueError("Family must be one of: gaussian, binomial, gamma, inverse_gaussian or poisson!")
+            raise ValueError(
+                "Family must be one of: gaussian, binomial, gamma, inverse_gaussian or poisson!")
         self.fitted = False
         self.formula = formula
         self.data = copy(data)
@@ -84,14 +86,14 @@ class Lmer(object):
 
     def __repr__(self):
         out = "{}.{}(fitted = {}, formula = {}, family = {})".format(
-        self.__class__.__module__,
-        self.__class__.__name__,
-        self.fitted,
-        self.formula,
-        self.family)
+            self.__class__.__module__,
+            self.__class__.__name__,
+            self.fitted,
+            self.formula,
+            self.family)
         return out
 
-    def _make_factors(self,factor_dict, ordered=False):
+    def _make_factors(self, factor_dict, ordered=False):
         """
         Covert specific columns to R-style factors. Default scheme is dummy coding where reference is 1st level provided. Alternative is orthogonal polynomial contrasts. User can also specific custom contrasts.
 
@@ -131,14 +133,14 @@ class Lmer(object):
             df[k] = df[k].astype(str)
 
         r_df = pandas2ri.py2ri(df)
-        for k,v in factor_dict.items():
+        for k, v in factor_dict.items():
             if isinstance(v, list):
-                r_df = factorize(r_df,k,v)
+                r_df = factorize(r_df, k, v)
             elif isinstance(v, dict):
                 levels = list(v.keys())
                 contrasts = np.array(list(v.values()))
-                r_df = factorize(r_df,k,levels)
-                r_df = contrastize(r_df,k,contrasts)
+                r_df = factorize(r_df, k, levels)
+                r_df = contrastize(r_df, k, contrasts)
         return r_df
 
     def _refit_orthogonal(self):
@@ -151,9 +153,10 @@ class Lmer(object):
         new_factors = {}
         for k in self.factors.keys():
             new_factors[k] = sorted(list(map(str, self.data[k].unique())))
-        self.fit(factors = new_factors,ordered=True,summarize=False,permute=self._permute,conf_int=self._conf_int,REML=self._REML)
+        self.fit(factors=new_factors, ordered=True, summarize=False,
+                 permute=self._permute, conf_int=self._conf_int, REML=self._REML)
 
-    def anova(self,force_orthogonal=False):
+    def anova(self, force_orthogonal=False):
         """
[…]
