+++ pymer4/simulate.py
@@ -4,14 +4,15 @@ import pandas as pd
 from scipy.spatial.distance import squareform
 from pymer4.utils import discrete_inverse_logit, isPSD, nearestPSD
 
-__all__  = ['easy_multivariate_normal',
-            'simulate_lm',
-            'simulate_lmm']
+__all__ = ['easy_multivariate_normal',
+           'simulate_lm',
+           'simulate_lmm']
 
 __author__ = ['Eshin Jolly']
 __license__ = "MIT"
 
-def simulate_lm(num_obs,num_coef,coef_vals=None,corrs=None,mus=0.0,sigmas=1.0,noise_params=(0,1),family='gaussian',seed=None):
+
+def simulate_lm(num_obs, num_coef, coef_vals=None, corrs=None, mus=0.0, sigmas=1.0, noise_params=(0, 1), family='gaussian', seed=None):
     """
     Function to quickly simulate a regression model dataset, with continuous predictors.
     Provided a number of observations, number of coefficients, and optionally correlations between predictors, means, and standard deviations of predictors, returns a pandas dataframe with simulated data that can be used to estimate a linear regression using Lm(). Using the family='binomial' argument can generate discrete dependent variable values for use with logistic regression.
@@ -42,37 +43,43 @@ def simulate_lm(num_obs,num_coef,coef_vals=None,corrs=None,mus=0.0,sigmas=1.0,no
 
     if coef_vals is not None:
         if len(coef_vals) - num_coef == 0:
-            raise ValueError("Missing one coefficient value. Did you provide a value for the intercept term?")
+            raise ValueError(
+                "Missing one coefficient value. Did you provide a value for the intercept term?")
         else:
-            assert len(coef_vals) == num_coef + 1, "Number of coefficient values should be num_coef + 1 (for intercept)"
+            assert len(coef_vals) == num_coef + \
+                1, "Number of coefficient values should be num_coef + 1 (for intercept)"
 
         b = coef_vals
     else:
-        b = np.random.rand(num_coef+1)
+        b = np.random.rand(num_coef + 1)
 
-    if isinstance(mus,list) or isinstance(mus,np.ndarray):
-        assert len(mus) == len(b)-1, "mus must match number of num_coef"
-    if isinstance(sigmas,list) or isinstance(sigmas,np.ndarray):
-        assert len(sigmas) == len(b)-1, "sigmas must match number of num_coef"
-    assert isinstance(noise_params,tuple) and len(noise_params) == 2, "noise_params should be a tuple of (mean,std)"
+    if isinstance(mus, list) or isinstance(mus, np.ndarray):
+        assert len(mus) == len(b) - 1, "mus must match number of num_coef"
+    if isinstance(sigmas, list) or isinstance(sigmas, np.ndarray):
+        assert len(sigmas) == len(b) - 1, "sigmas must match number of num_coef"
+    assert isinstance(noise_params, tuple) and len(
+        noise_params) == 2, "noise_params should be a tuple of (mean,std)"
 
     # Generate random design matrix
     if corrs is not None:
-        X = easy_multivariate_normal(num_obs,num_coef,corrs,mus,sigmas,seed)
+        X = easy_multivariate_normal(
+            num_obs, num_coef, corrs, mus, sigmas, seed)
     else:
-        X = np.random.normal(mus,sigmas,size=(num_obs,num_coef))
+        X = np.random.normal(mus, sigmas, size=(num_obs, num_coef))
     # Add intercept
-    X = np.column_stack([np.ones((num_obs,1)),X])
+    X = np.column_stack([np.ones((num_obs, 1)), X])
     # Generate data
[…]

[…]
