### 2018-08-20 #862 @ejolly
@cMadan I believe I have addressed all the issues brought up so far, but there's one in particular (https://github.com/ejolly/pymer4/issues/28) opened by @janfreyberg, and related to the installation issues @RMHogervorst was running into, that I would really appreciate some advice on. In summary, I'm not sure whether to favor support for both Python 2 and 3, or drop Python 2 support but make overall installation a bit easier for some users. Rather than reiterating it all here, I've noted my question in that issue thread. Thanks!
