### 2018-08-21 0b7bbde
Edited travis file to drop Python 2 environment and update R package versions.
