### 2018-08-22 804d742
Updated rpy2 dependency version. Freeze pytest version on travis. Free pytest version on tox.
