+++ requirements.txt
@@ -1,6 +1,6 @@
 pandas>=0.19.1
 numpy>=1.12.0
-rpy2==2.8.5
+rpy2>=2.9.4
 seaborn>=0.8.0
 matplotlib>=2.0
 patsy>=0.4.1
