+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
 envlist = py27,py36
 [testenv]
-deps=pytest
+deps=pytest==3.6.1
 commands=pytest
