+++ .travis.yml
@@ -6,6 +6,7 @@ python:
 install:
 - sudo pip install -r requirements.txt
 - sudo python setup.py install
+- sudo pip install pytest==3.6.1
 script:
 - sudo py.test
 before_install:
