### 2018-08-23 c4466ae
Fixed tox environment. Now only tests python 3. Trying new travis setup.
