+++ tox.ini
@@ -1,5 +1,6 @@
 [tox]
-envlist = py27,py36
+envlist = py36
 [testenv]
-deps=pytest==3.6.1
+deps= pytest==3.6.1
+      tzlocal
 commands=pytest
