### 2018-08-23 c93d7a1
Removed root init to fix pytest issues for tox and CI.
