+++ .travis.yml
@@ -18,4 +18,4 @@ before_install:
 - sudo apt-get -y install r-base r-base-dev python-tk libnlopt-dev
 - sudo R -e "install.packages('versions', repos = 'http://cran.us.r-project.org');
   library(versions); install.versions(c('lme4','lmerTest','lsmeans'),c('1.1-17','3.0-1','2.27-62'))"
-- sudo pip install pytest==3.6.1
+- sudo pip install pytest
