+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 envlist = py36
 [testenv]
-deps= pytest==3.6.1
+deps= pytest
       tzlocal
 commands=pytest
