### 2018-08-23 40adaed
Fix formatting in travis file.
Clean up tests.

Fix travis file.

Fix travis file.
