### 2018-08-23 8e789b8
Fix formatting in travis file.
