### 2018-08-23 c32d1e3
Clean up tests.
