+++ pymer4/tests/test_models.py
@@ -134,11 +134,6 @@ def test_poisson_lmm():
     m = Lmer('DV_int ~ IV3 + (1|Group)', data=df, family='poisson')
     m.fit(summarize=False)
     assert m.family == 'poisson'
-    # res = np.array([[ 1.64640281e+000,  1.54854277e+000,  1.74426286e+000,
-    #      4.99295118e-002,  3.29745426e+001,  1.88245501e-238],
-    #    [-5.40993372e-002, -1.45285282e-001,  3.70866072e-002,
-    #      4.65242959e-002, -1.16281904e+000,  2.44902930e-001]])
-    # assert np.allclose(m.coefs.iloc[:,:-1].values, res)
     assert m.coefs.shape == (2, 7)
     assert 'Z-stat' in m.coefs.columns
 
@@ -151,11 +146,6 @@ def test_gamma_lmm():
     m = Lmer('DV_g ~ IV3 + (1|Group)', data=df, family='gamma')
     m.fit(summarize=False)
     assert m.family == 'gamma'
-    # res = np.array([[ 6.73406340e-01,  6.42966702e-01,  7.03845977e-01,
-    #      1.55307127e-02,  4.33596548e+01,  0.00000000e+00],
-    #    [-6.90225926e-03, -3.33536803e-02,  1.95491618e-02,
-    #      1.34958710e-02, -5.11434888e-01,  6.09046570e-01]])
-    # assert np.allclose(m.coefs.iloc[:,:-1].values,res)
     assert m.coefs.shape == (2, 7)
 
 
@@ -167,9 +157,4 @@ def test_inverse_gaussian_lmm():
     m = Lmer('DV_g ~ IV3 + (1|Group)', data=df, family='inverse_gaussian')
     m.fit(summarize=False)
     assert m.family == 'inverse_gaussian'
-    # res = np.array([[ 4.54086509e-001,  4.12691749e-001,  4.95481268e-001,
-    #      2.11201633e-002,  2.15001419e+001,  1.55212497e-102],
-    #    [-9.15658032e-003, -4.51946155e-002,  2.68814549e-002,
-    #      1.83870905e-002, -4.97989625e-001,  6.18491355e-001]])
-    # assert np.allclose(m.coefs.iloc[:,:-1].values,res)
     assert m.coefs.shape == (2, 7)
