+++ .travis.yml
@@ -2,12 +2,12 @@ sudo: true
 dist: trusty
 langauge: python
 python:
-- "3.6"
+- '3.6'
 install:
 - sudo pip install -r requirements.txt
 - sudo python setup.py install
 script:
-  - sudo py.test
+- sudo py.test
 before_install:
 - sudo sh -c 'echo "deb http://cran.rstudio.com/bin/linux/ubuntu trusty/" >> /etc/apt/sources.list'
 - gpg --keyserver keyserver.ubuntu.com --recv-key E084DAB9
