+++ .travis.yml
@@ -2,6 +2,7 @@ sudo: true
 dist: trusty
 langauge: python
 python:
+- "2.7"
 - "3.6"
 install:
 - sudo pip install -r requirements.txt
