### 2018-08-27 #862 @cMadan
While supporting both Python 2 and 3 is admirable, at some point you will have to drop Python 2 support since that's inevitable and happing fast (e.g., see https://python3statement.org). If there are issues currently in trying to remain cross-compatible, my inclination is to stick with just Python 3 already--and then streamline the install experience for that procedure. Ultimately the decision is yours of course, but that's the way I see it.
