### 2018-08-27 b49299e
Updated travis config to use conda-forge for rpy2 and r installs. Made some tests less intensive because conda-forge installs are slower.
