### 2018-08-27 dd6a004
Edit travis file to hardcode python 3.
