+++ .travis.yml
@@ -1,17 +1,14 @@
-sudo: false
 langauge: python
-python:
-  - "3.6"
-
+python: "3.6"
 before_install:
-  - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
+  - wget http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
   - bash miniconda.sh -b -p $HOME/miniconda
   - export PATH="$HOME/miniconda/bin:$PATH"
   - hash -r
   - conda config --set always_yes yes --set changeps1 no
   - conda update conda
 install:
-- conda create -n testenv python=$TRAVIS_PYTHON_VERSION pip pytest
+- conda create -n testenv python=3.6 pip pytest
 - source activate testenv
 - conda install -c conda-forge rpy2 r-lmertest r-lsmeans
 - pip install -r requirements.txt
