### 2018-08-27 daeda01
Updated installation instructions and added quick installation test function.
