+++ pymer4/version.py
@@ -1,4 +1,4 @@
 """Specifies current version of pymer4 to be used by setup.py and __init__.py
 """
 
-__version__ = '0.5.0'
+__version__ = '0.6.0'
