+++ docs/index.rst
@@ -48,23 +48,24 @@ This package has some extra goodies to make life a bit easier, namely:
 
 Installation
 ------------
-Requires a working installation of *both* Python (2.7 or 3.6) and R (>= 3.2.4).
+:code:`pymer4` since version 0.6.0 is only compatible with Python 3. Versions 0.5.0 and lower will work with Python 2, but will not contain any new features. :code:`pymer4` also requires a working R installation with specific packages installed and it will *not* install R or these packages for you. However, you can follow either option below to easily handle these dependencies.
 
-You will also need the :code:`lme4`, :code:`lmerTest`, and :code:`lsmeans` R packages installed.
+Option 1 (simpler but slower model fitting)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+If you don't have R installed and you use the Anaconda Python distribution simply run the following commands to have Anaconda install R and the required packages for you. This is fairly painless installation, but model fitting will be slower than if you install R and ``pymer4`` separately and configure them (option 2).
 
-*This package will not install R or R packages for you!*
+1. ``conda install -c conda-forge rpy2 r-lmertest r-lsmeans``  
+2. ``pip install pymer4``
 
-1. Method 1 - Install from PyPi (stable)
+Option 2 (potentially trickier, but faster model fitting)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+This method assumes you already have R installed. If not install first install it from the `R Project website <https://www.r-project.org/>`_. Then complete the following steps: 
 
-.. code-block:: python
+1. Install the required R packages by running the following command from within R: ``install.packages(c('lme4','lmerTest','lsmeans'))``
+2. Install pymer4 by running the following command in a terminal: ``pip install pymer4``
+3. Test the installation to see if it's working by running the following command in a terminal: ``python -c "from pymer4.test_install import test_install; test_install()"``  
 
-    pip install pymer4
-
-2. Method 2 - Install from github (latest)
-
-.. code-block:: python
-
-    pip install git+https://github.com/ejolly/pymer4
+If this produces any errors they are likely related to ``rpy2``. See the section below for a fix.
 
 Install issues
 ^^^^^^^^^^^^^^
@@ -78,18 +79,18 @@ Some users have issues installing ``pymer4`` on recent versions of macOS. This i
 
     .. code-block:: bash
 
-        export CC="$(find `brew info gcc | grep usr | sed 's/(.*//' | awk '{printf $1"/bin"}'` -name 'x86*gcc-7')"
+        export CC="$(find `brew info gcc | grep usr | sed 's/(.*//' | awk '{printf $1"/bin"}'` -name 'x86*gcc-?')"
         export CFLAGS="-W"
 
-6. If the above results in any error output (it should return nothing) you might need to manually find out where the new compiler is installed. To do so use ``brew info gcc`` and ``cd`` into the directory that begins with ``/usr`` in the output of that command. From there ``cd`` into ``bin`` and look for a file that begins with ``x86`` and ends with ``gcc-7``. Copy the *full path* to that file and run the following:
[…]
