+++ .travis.yml
@@ -8,7 +8,7 @@ install:
 - pip install -r requirements.txt
 - python setup.py install
 script:
-- sudo py.test
+- py.test
 before_install:
   - wget http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
   - bash miniconda.sh -b -p $HOME/miniconda
