### 2018-10-19 74b1755
Fixed bug where Lm models failed to summarize if some parameters were 0 and t-stats were nan.
