### 2018-11-18 #862 @cMadan
@ejolly, how are things going?
