### 2018-11-19 cf8ae71
Updated installation instructions.
