+++ README.md
@@ -15,45 +15,48 @@ Current documentation and usage examples can be found **[here](http://eshinjolly
 
 ## Installation  
 
-`pymer4` since version 0.6.0 is only compatible with Python 3. Versions 0.5.0 and lower will work with Python 2, but will not contain any new features. `pymer4` also requires a working R installation with specific packages installed and it will *not* install R or these packages for you. However, you can follow either option below to easily handle these dependencies. 
+`pymer4` since version 0.6.0 is only compatible with Python 3. Versions 0.5.0 and lower will work with Python 2, but will not contain any new features. `pymer4` also requires a working R installation with specific packages installed and it will *not* install R or these packages for you. However, you can follow either option below to easily handle these dependencies.
 
 ## Option 1 (simpler but slower model fitting)
 
 If you don't have R installed and you use the Anaconda Python distribution simply run the following commands to have Anaconda install R and the required packages for you. This is fairly painless installation, but model fitting will be slower than if you install R and `pymer4` separately and configure them (option 2).
 
-1. `conda install -c conda-forge rpy2 r-lmertest r-lsmeans`  
+1. `conda install -c conda-forge r r-base r-lmertest r-lsmeans rpy2`  
 2. `pip install pymer4`
+3. Test the installation to see if it's working by running: `python -c "from pymer4.test_install import test_install; test_install()"`
+4. If there are errors follow the guide below
 
 ## Option 2 (potentially trickier, but faster model fitting)  
 
-This method assumes you already have R installed. If not install first install it from the [R Project website](https://www.r-project.org/). Then complete the following steps: 
+This method assumes you already have R installed. If not install first install it from the [R Project website](https://www.r-project.org/). Then complete the following steps:
 
 1. Install the required R packages by running the following command from within R: `install.packages(c('lme4','lmerTest','lsmeans'))`
 2. Install pymer4: `pip install pymer4`
-3. Test the installation to see if it's working by running: `python -c "from pymer4.test_install import test_install; test_install()"`  
-
-If this produces any errors they are likely related to `rpy2`. See the the section below for a fix. 
+3. Test the installation to see if it's working by running: `python -c "from pymer4.test_install import test_install; test_install()"`
+4. If there are errors follow the guide below  
 
 ### Installation issues
 
-If you run into issues using Option 2 above, it's likely due to compiler issues that give `rpy2` (a package dependency of `pymer4`) some issues during install. The instructions below should fix that on macOS:
+If you run into issues using either option above, it's likely due to compiler issues that give `rpy2` (a package dependency of `pymer4`) some issues during install. The instructions below should fix that on macOS:
 
 1. Install [homebrew](https://brew.sh/) if you don't have it already, by running the command at the link (it's a great pacakage manager for macOS). To check if you already have it, do `which brew` in your Terminal. If nothing pops up you don't have it.
 2. Fix brew permissions: `sudo chown -R $(whoami) $(brew --prefix)/*` (this is **necessary** on macOS Sierra or higher (>= macOS 10.12))
 3. Update homebrew `brew update`
-4. Install an updated compiler: `brew install gcc`, or if you have homebrew already, `brew upgrade gcc`
-5. Enable the new compiler for use:
+4. Install the xz utility `brew install xz`
+5. At this point you can try to re-install `pymer4` and re-test the installation. If it still doesn't work follow the next few steps below
+6. Install an updated compiler: `brew install gcc`, or if you have homebrew already, `brew upgrade gcc`
+7. Enable the new compiler for use:
     ```
     export CC="$(find `brew info gcc | grep usr | sed 's/(.*//' | awk '{printf $1"/bin"}'` -name 'x86*gcc-?')"
     export CFLAGS="-W"
     ```
[…]
