### 2018-11-19 786872b
Edited installation instructions from easy_install branch with some modifications.
