### 2018-11-19 #862 @ejolly
Hey @cMadan sorry this has a been a bit frustrating to sort out various installation issues but I think thinks should be set for the most part now! The trickiness has really been with `rpy2` a core dependency of `pymer4`. Despite much effort from their development team it appears to me that it's relatively well known in the community that installation of that pacakage can be tricky.

I just pushed a new version to pypi and thoroughly updated the README and documentation site with two different installation options. The first should be much easier for most Python users who already use the Anaconda distribution. This was suggested by @janfreyberg 's review in https://github.com/ejolly/pymer4/issues/28. I've also expanded some instructions for an alternative installation method for R or Python users that don't use Anaconda.  I've done the best I can to support the most common issues that users may encounter and have tested both installation methods myself on numerous machines. As per your suggestion "dropping" Python 2 support in this new version helped a bit; in many cases Python 2 _should_ work I'm just not adding official support for it in versions >= 0.6.0. Users still can install 0.5.0 if they are insistent on Python 2.

Please let me know if there's anything else I need to do for the review process!
