### 2018-11-19 #862 @janfreyberg
This looks good to me and the new installation instructions are very good! I like the `test_install()` function, too.
