### 2018-11-23 #862 @cMadan
@janfreyberg, thank you for your thorough review, it is much appreciate!

@RMHogervorst, I am going to remove you as an assigned reviewer here, but I appreciate your feedback with the installation issues earlier on. Hopefully the changes to the project have now ameliorated the issues you were having.

@ejolly, I will do a final look over of the project in the next few days and get back to you.
