### 2018-11-25 #862 @cMadan
@ejolly, everything looks good, just two things to do:
- mint a release for 0.6.0 and tell me the Zenodo DOI for the archived code
