### 2018-11-26 88bce73
Updated zenodo DOI and pypi badge with 0.6.0 release.
