### 2018-11-26 #862 @ejolly
@cMadan Done and done. The readme contains the correct badge link for both. Here is the [latest DOI](https://zenodo.org/record/1523205). Please let me know if you need anything else!
