### 2018-11-26 #862 @cMadan
@ejolly, looks like your bib is missing a comma at the end of line 1.

Otherwise you're all set for the submission to be accepted!
