+++ paper.bib
@@ -1,4 +1,4 @@
-@Article{Bates:2015
+@Article{Bates:2015,
    title = {Fitting Linear Mixed-Effects Models Using {lme4}},
    author = {Douglas Bates and Martin M{\"a}chler and Ben Bolker and
      Steve Walker},
