### 2018-11-26 9347acc
Added JOSS badge and citation information.
