### 2019-01-18 ab8e284
Adding save/load/Rlike descriptions for fixef
This version adds a save/load routing to the lmer model. 
Models can be saved by model.save(filename, filedirectory)
and loaded by model.load(filename, filedirectory).

The save routine dumps the model's files to fname.[pkl/robj/csv] in the directory load_d 
To split the model in three files is necessary. Simply using pickle would 
lead to models with big training data, because pickle cannot efficently save 
self.model_obj or panda DataFrames. This leads to the problem, that after saving
the model cannot be loaded anymore, because the unpacking nee
[…]
